# Avoid allocations, fix CI