--- a/standardtracer/span.go
+++ b/standardtracer/span.go
@@ -11,22 +11,27 @@ import (
 // Implements the `Span` interface. Created via tracerImpl (see
 // `standardtracer.New()`).
 type spanImpl struct {
-	lock     sync.Mutex
-	tracer   *tracerImpl
-	recorder SpanRecorder
-	raw      RawSpan
+	tracer     *tracerImpl
+	sync.Mutex // protects the fields below
+	raw        RawSpan
+	// TODO(tschottdorf): should this be available to the Recorder
+	// via RawSpan as well?
+	traceAttrs map[string]string // initialized on first use
 }
 
 func (s *spanImpl) SetOperationName(operationName string) opentracing.Span {
-	s.lock.Lock()
-	defer s.lock.Unlock()
+	s.Lock()
+	defer s.Unlock()
 	s.raw.Operation = operationName
 	return s
 }
 
 func (s *spanImpl) SetTag(key string, value interface{}) opentracing.Span {
-	s.lock.Lock()
-	defer s.lock.Unlock()
+	s.Lock()
+	defer s.Unlock()
+	if s.raw.Tags == nil {
+		s.raw.Tags = opentracing.Tags{}
+	}
 	s.raw.Tags[key] = value
 	return s
 }
@@ -45,14 +50,14 @@ func (s *spanImpl) LogEventWithPayload(event string, payload interface{}) {
 }
 
 func (s *spanImpl) Log(ld opentracing.LogData) {
-	s.lock.Lock()
-	defer s.lock.Unlock()
+	s.Lock()
+	defer s.Unlock()
 
 	if ld.Timestamp.IsZero() {
 		ld.Timestamp = time.Now()
 	}
 
-	s.raw.Logs = append(s.raw.Logs, &ld)
+	s.raw.Logs = append(s.raw.Logs, ld)
 }
 
 func (s *spanImpl) Finish() {
@@ -66,13 +71,13 @@ func (s *spanImpl) FinishWithOptions(opts opentracing.FinishOptions) {
 	}
 	duration := finishTime.Sub(s.raw.Start)
 
-	s.lock.Lock()
-	defer s.lock.Unlock()
+	s.Lock()
+	defer s.Unlock()
 	if opts.BulkLogData != nil {
 		s.raw.Logs = append(s.raw.Logs, opts.BulkLogData...)
 	}
 	s.raw.Duration = duration
-	s.recorder.RecordSpan(&s.raw)
+	s.tracer.recorder.RecordSpan(s.raw)
 }
 
 func (s *spanImpl) SetTraceAttribute(restrictedKey, val string) opentracing.Span {
@@ -81,10 +86,12 @@ func (s *spanImpl) SetTraceAttribute(restrictedKey, val string) opentracing.Span
 		panic(fmt.Errorf("Invalid key: %q", restrictedKey))
 	}
 
-	s.raw.StandardContext.attrMu.Lock()
-	defer s.raw.StandardContext.attrMu.Unlock()
-
-	s.raw.StandardContext.traceAttrs[canonicalKey] = val
+	s.Lock()
+	defer s.Unlock()
+	if s.traceAttrs == nil {
+		s.traceAttrs = make(map[string]string)
+	}
+	s.traceAttrs[canonicalKey] = val
 	return s
 }
 
@@ -94,10 +101,10 @@ func (s *spanImpl) TraceAttribute(restrictedKey string) string {
 		panic(fmt.Errorf("Invalid key: %q", restrictedKey))
 	}
 
-	s.raw.StandardContext.attrMu.RLock()
-	defer s.raw.StandardContext.attrMu.RUnlock()
+	s.Lock()
+	defer s.Unlock()
 
-	return s.raw.StandardContext.traceAttrs[canonicalKey]
+	return s.traceAttrs[canonicalKey]
 }
 
 func (s *spanImpl) Tracer() opentracing.Tracer {