--- a/standardtracer/util.go
+++ b/standardtracer/util.go
@@ -7,13 +7,19 @@ import (
 )
 
 var (
-	seededIDGen  = rand.New(rand.NewSource(time.Now().UnixNano()))
+	seededIDGen = rand.New(rand.NewSource(time.Now().UnixNano()))
+	// The golang rand generators are *not* intrinsically thread-safe.
 	seededIDLock sync.Mutex
 )
 
 func randomID() int64 {
-	// The golang rand generators are *not* intrinsically thread-safe.
 	seededIDLock.Lock()
 	defer seededIDLock.Unlock()
 	return seededIDGen.Int63()
+}
+
+func randomID2() (int64, int64) {
+	seededIDLock.Lock()
+	defer seededIDLock.Unlock()
+	return seededIDGen.Int63(), seededIDGen.Int63()
 }