Embed StandardContext (not *StandardContext) into
RawSpan and initialize the attribute and tags map
only on the first write (it is valid to read from
a nil map).

Move the trace attributes from StandardContext onto
spanImpl; they were not exposed to the recorder, and
the lock on StandardContext required that object to
always be heap allocated. If they need to be exposed
to the user, they can be moved onto RawSpan (or even
back onto StandardContext, but without the lock).
Locking is instead synchronized through the Span's
log.

Work towards sane allocation behavior and add some
benchmarks to that effect: Span creation and destruction
with calls to LogEvent, SetTag oder SetTraceAttribute
now don't allocate.

Changed various other pointers to be passed by-value.

Fix various linter errors and the invocation in
the Makefile (which did not exit nonzero on issues).

Use a sync.Pool instead of new allocations …
This reduces the allocs for a common Span to zero.

```
name               old time/op    new time/op    delta
Span_Empty-8          411ns ± 0%     337ns ± 0%   ~     (p=1.000 n=1+1)
Span_100Events-8     18.2µs ± 0%    18.6µs ± 0%   ~     (p=1.000 n=1+1)
Span_1000Events-8    18.2µs ± 0%    19.3µs ± 0%   ~     (p=1.000 n=1+1)
Span_100Tags-8       25.1µs ± 0%    26.2µs ± 0%   ~     (p=1.000 n=1+1)
Span_1000Tags-8      24.7µs ± 0%    24.9µs ± 0%   ~     (p=1.000 n=1+1)

name               old alloc/op   new alloc/op   delta
Span_Empty-8           144B ± 0%       0B ±NaN%   ~     (p=1.000 n=1+1)
Span_100Events-8     16.2kB ± 0%    16.1kB ± 0%   ~     (p=1.000 n=1+1)
Span_1000Events-8    16.2kB ± 0%    16.1kB ± 0%   ~     (p=1.000 n=1+1)
Span_100Tags-8       10.7kB ± 0%    10.5kB ± 0%   ~     (p=1.000 n=1+1)
Span_1000Tags-8      10.7kB ± 0%    10.5kB ± 0%   ~     (p=1.000 n=1+1)

name               old allocs/op  new allocs/op  delta
Span_Empty-8           1.00 ± 0%     0.00 ±NaN%   ~     (p=1.000 n=1+1)
Span_100Events-8       9.00 ± 0%      8.00 ± 0%   ~     (p=1.000 n=1+1)
Span_1000Events-8      9.00 ± 0%      8.00 ± 0%   ~     (p=1.000 n=1+1)
Span_100Tags-8         12.0 ± 0%      11.0 ± 0%   ~     (p=1.000 n=1+1)
Span_1000Tags-8        12.0 ± 0%      11.0 ± 0%   ~     (p=1.000 n=1+1)
```

FYI, I'm piling more commits on top of this, but would be good to get this
in/discuss it before it gets more refined.