@bensigelman 2016-02-14
@tschottdorf the standardtracer stuff is not optimized (or at least it was not optimized) and I appreciate you taking the time to profile and fix it up. https://github.com/opentracing/opentracing-go/pull/53#discussion-diff-52850024 is the only concern I have... I wish we had a more realistic real-world benchmark. I'm okay proceeding with the non-pointer `[]LogData`, though, unless @yurishkuro or others (@slimsag?) with appreciable Go experience have words of warning to share on this front.

Basically LGTM, though, and thanks again!