@tschottdorf 2016-02-14
> I'm okay proceeding with the non-pointer []LogData, though, unless @yurishkuro or others (@slimsag?) with appreciable Go experience have words of warning to share on this front.

I'd be surprised if the pointer were better (we could always change it back). It's definitely worse in that unless you're tricky, each `LogEntry` is a single allocation in the previous version.