@yurishkuro 2016-02-14
lgtm

I would suggest merging incrementally than piling more commits onto this pr.